% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneuploidy_zscores_pascores.R
\name{countAndNormalize}
\alias{countAndNormalize}
\title{Counts number of reads overlapping bins and performs GC-correction}
\usage{
countAndNormalize(bins, measure)
}
\arguments{
\item{bins}{a \code{GRanges} object, obtained from \function{makeBins}}

\item{reads}{a \code{GRanges} object, obtained from \function{filterReads}}
}
\value{
a \code{GRanges} object, with columns representing the raw counts
 for each bin (counts), the log2 weighted count (counts.mult),
 the predicted log2 coverage based on LOESS (loess.pred), and the
 normalized coverage (adjusted). The adjusted column is used for further
 z-score and PA-score analysis
}
\description{
Uses LOESS to return GC-adjusted read-depth for desired bins
}
