% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneuploidy_zscores_pascores.R
\name{getPAscores}
\alias{getPAscores}
\title{Aneuploidy PA-scores}
\usage{
getPAscores(zscores, loo = FALSE)
}
\arguments{
\item{loo}{a logical scalar, indicating whether the PA-scores for the normal samples
should be calculated leave-one-out, as described in the paper. Default is
FALSE}

\item{bincounts.list}{a list, with each object being a \code{GRanges}
object obtained from \function{countAndNormalize}}

\item{normal.indices}{a integer vector with the indices specifying
which elements in bincounts.list are to be used as controls}
}
\value{
a vector with the PA-scores for each sample
}
\description{
Returns the PA-scores based on normalized genome coverage.
}
