% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneuploidy_zscores_pascores.R
\name{getZscores}
\alias{getZscores}
\title{Aneuploidy z-scores}
\usage{
getZscores(
  bins,
  normal.ids = NULL,
  loo = FALSE,
  refpanel = NULL,
  measure = "cov"
)
}
\arguments{
\item{loo}{logical, indicating whether the z-scores for the normal samples
should be calculated leave-one-out, as described in the paper. Default is
FALSE}

\item{bincounts.list}{a list, with each object being a \code{GRanges}
object obtained from \function{countAndNormalize}}

\item{normal.indices}{a integer vector with the indices specifying
which elements in bincounts.list are to be used as controls}
}
\value{
a list, with each member being a numeric vector with the arm-level
 z-scores for that sample
}
\description{
Returns the z-scores based on normalized genome coverage.
}
